refactor: 투표 리스트 응답값 공통 타입으로 분리
